Factor out scheduling logic from job computation classes
