+++ pyproject.toml
@@ -134,11 +134,13 @@ packages = [
     "spatialprofilingtoolbox.entry_point",
     "spatialprofilingtoolbox.apiserver",
     "spatialprofilingtoolbox.apiserver.app",
+    "spatialprofilingtoolbox.apiserver.request_scheduling",
     "spatialprofilingtoolbox.apiserver.scripts",
     "spatialprofilingtoolbox.graphs",
     "spatialprofilingtoolbox.graphs.scripts",
     "spatialprofilingtoolbox.ondemand",
     "spatialprofilingtoolbox.ondemand.providers",
+    "spatialprofilingtoolbox.ondemand.computers",
     "spatialprofilingtoolbox.ondemand.scripts",
     "spatialprofilingtoolbox.db",
     "spatialprofilingtoolbox.db.accessors",
