+++ spatialprofilingtoolbox/apiserver/request_scheduling/counts_scheduler.py
@@ -1,105 +1,18 @@
-"""Count cells for a specific signature, over the specially-created binary-format index."""
 
 from typing import cast
 
-from numpy import sum
-from numpy import uint64 as np_int64
-from numpy.typing import NDArray
-
-from spatialprofilingtoolbox.db.exchange_data_formats.metrics import PhenotypeCriteria
-from spatialprofilingtoolbox.ondemand.providers.provider import CellDataArrays
-from spatialprofilingtoolbox.ondemand.providers.pending_provider import PendingProvider
-from spatialprofilingtoolbox.ondemand.phenotype_str import (
-    phenotype_str_to_phenotype,
-    phenotype_to_phenotype_str,
-)
-from spatialprofilingtoolbox.db.describe_features import get_feature_description
 from spatialprofilingtoolbox.db.database_connection import DBCursor
+from spatialprofilingtoolbox.db.exchange_data_formats.metrics import PhenotypeCriteria
+from spatialprofilingtoolbox.ondemand.relevant_specimens import retrieve_cells_selected
+from spatialprofilingtoolbox.ondemand.phenotype_str import phenotype_to_phenotype_str
+from spatialprofilingtoolbox.db.describe_features import get_feature_description
+from spatialprofilingtoolbox.apiserver.request_scheduling.computation_scheduler import GenericComputationScheduler
 from spatialprofilingtoolbox.standalone_utilities.log_formats import colorized_logger
-ItemList = tuple[str, list[int]]
-Item = tuple[str, tuple[int, ...]]
 
 logger = colorized_logger(__name__)
 
 
-class CountsProvider(PendingProvider):
-    """Scan binary-format expression matrices for specific signatures."""
-
-    def compute(self) -> None:
-        args, arrays = self._prepare_parameters()
-        if arrays.identifiers is None:
-            self.handle_insert_value(None, allow_null=True)
-        else:
-            count = self._perform_count(args, arrays)
-            self.handle_insert_value(count)
-
-    def _prepare_parameters(self) -> tuple[tuple[int, int, tuple[int, ...]], CellDataArrays]:
-        study = self.job.study
-        specification = str(self.job.feature_specification)
-        _, specifiers = self.retrieve_specifiers(study, specification)
-        phenotype = phenotype_str_to_phenotype(specifiers[0])
-        cells_selected = self._retrieve_cells_selected(study, specification)  # to deprecate
-        marker_set = (phenotype.positive_markers, phenotype.negative_markers)
-        arrays = self.get_cell_data_arrays()
-        features = tuple(n.symbol for n in arrays.feature_names.names)
-        signatures = tuple(map(lambda m: cast(int, self._compute_signature(m, features)), marker_set))
-        return ((signatures[0], signatures[1], cells_selected), arrays)
-
-    @staticmethod
-    def _retrieve_cells_selected(study: str, specification: str) -> tuple[int, ...]:
-        with DBCursor(study=study) as cursor:
-            query = 'SELECT histological_structure FROM cell_set_cache WHERE feature=%s ;'
-            cursor.execute(query, (specification,))
-            rows = tuple(cursor.fetchall())
-        if len(rows) == 0:
-            return ()
-        return tuple(map(lambda row: int(row[0]), rows))
-
-    @staticmethod
-    def _perform_count(args: tuple[int, int, tuple[int, ...]], arrays: CellDataArrays) -> int:
-        if arrays.identifiers.shape[0] == 0:
-            return 0
-        return CountsProvider._count_structures_of_partial_signed_signature(*args, arrays)
-
-    @staticmethod
-    def _count_structures_of_partial_signed_signature(
-        positives_signature: int,
-        negatives_signature: int,
-        cells_selected: tuple[int, ...],
-        arrays: CellDataArrays,
-    ) -> int:
-        """Count the number of cells in the given sample that match this signature."""
-        if positives_signature == 0 and negatives_signature == 0:
-            return arrays.identifiers.shape[0]
-        count = CountsProvider._get_count(arrays.phenotype, positives_signature, negatives_signature)
-        return count
-
-    @staticmethod
-    def _get_count(array_phenotype: NDArray[np_int64], positives_mask: int, negatives_mask: int) -> int:
-        """
-        Counts the number of elements of the list of integer-represented binary numbers which equal
-        to 1 along the bits indicated by the positives mask, and equal to 0 along the bits indicated
-        by the negatives mask.
-        """
-        return sum((array_phenotype | positives_mask == array_phenotype) &
-                   (~array_phenotype | negatives_mask == ~array_phenotype))
-
-    @staticmethod
-    def _compute_signature(
-        channel_names: tuple[str, ...],
-        all_features: tuple[str, ...],
-    ) -> int:
-        """Compute int signature of this channel name combination."""
-        missing = set(channel_names).difference(all_features)
-        if len(missing) > 0:
-            message = f'Cannot compute signature when these columns are requested: {missing}'
-            raise ValueError(message)
-        signature = 0
-        indices = map(lambda name: all_features.index(name), channel_names)
-        for index in indices:
-            signature = signature + (1 << index)
-        return signature
-
+class CountsScheduler(GenericComputationScheduler):
     @classmethod
     def get_or_create_feature_specification(
         cls,
@@ -120,7 +33,7 @@ class CountsProvider(PendingProvider):
         )
         specification = cls._get_feature_specification(study, *specifiers_arguments)
         if specification is not None:
-            _cells_selected = cls._retrieve_cells_selected(study, specification)
+            _cells_selected = retrieve_cells_selected(study, specification)
             if set(cells_selected) == set(_cells_selected):
                 return (specification, False)
         short = str(cells_selected[0:min(len(cells_selected), 5)]) + ' ...'
